--- a/src/GenSymIO.chpl
+++ b/src/GenSymIO.chpl
@@ -36,17 +36,17 @@ module GenSymIO {
      * Creates a pdarray server-side and returns the SymTab name used to
      * retrieve the pdarray from the SymTab.
      */
-    proc arrayMsg(cmd: string, payload: bytes, st: borrowed SymTab): MsgTuple throws {
+    proc arrayMsg(cmd: string, args: string, ref data: bytes, st: borrowed SymTab): MsgTuple throws {
         // Set up our return items
         var msgType = MsgType.NORMAL;
         var msg:string = "";
         var rname:string = "";
 
-        var (dtypeBytes, sizeBytes, data) = payload.splitMsgToTuple(b" ", 3);
+        var (dtypeBytes, sizeBytes) = args.splitMsgToTuple(" ", 2);
         var dtype = DType.UNDEF;
         var size:int;
         try {
-            dtype = str2dtype(dtypeBytes.decode());
+            dtype = str2dtype(dtypeBytes);
             size = sizeBytes:int;
         } catch {
             var errorMsg = "Error parsing/decoding either dtypeBytes or size";