@ronawho 2021-07-22
Probably easiest to review commit at a time. 

This optimizes out most unnecessary copies. There are some remaining in the Chapel ZMQ shim that we're looking at optimizing out, but on the Arkouda side the only "extra" copy is for to_ndarray when the memoryview returned to us is readonly (which is will be true for most pyzmq backends) at https://github.com/Bears-R-Us/arkouda/pull/885/files#diff-552f96acc4e4316d5d653916e1902f8f15ef77473e864e605c975aee228bbbb1R867. I plan on opening an issue against pyzmq to see if the view has to be readonly, but an alternative to this would be to make `to_ndarray()` return a readonly array and require users to create a copy if they want to modify it.