@ronawho 2021-07-23
> I didn't know about the memoryview stuff, it looks pretty cool 

Agreed, this was the first I learned of it too. These are some of the links I stashed away that I found the most helpful:
 - https://docs.python.org/3/c-api/buffer.html
 - https://www.python.org/dev/peps/pep-3118/
 - https://julien.danjou.info/high-performance-in-python-with-zero-copy-and-the-buffer-protocol/
 - https://stackoverflow.com/questions/18655648/what-exactly-is-the-point-of-memoryview-in-python
 - https://pyzmq.readthedocs.io/en/latest/serialization.html